refactor: Phase 3C - main execution test関数をmain_execution_test.goに抽出
Phase 3C完了によりmain_test.goの整理を完了。
main執行テスト関数を専用ファイルに分離し、ファイルサイズと保守性を改善。

## 変更内容
- cmd/beaver/main_execution_test.go (新規作成, 342行)
  - TestMainFunction: main()関数実行テスト
  - TestCommandStructure: コマンド登録・構造テスト
  - TestMainExecutionPaths: 実行パステスト
  - TestMainLogic: mainLogic関数直接テスト
  - TestCommandExecution: コマンド実行テスト

- cmd/beaver/main_test.go (1,195→865行, 28%削減)
  - 5個のmain実行テスト関数を削除
  - 出力テスト、ヘルパーテスト、境界条件テストのみ残存

## Phase 3完全成果
- Phase 3A+3B+3C総計: 76%削減 (3,550→865行)
- 論理的分離: 5個の専用テストファイル作成
- 全テスト機能・カバレッジ維持
- 保守性とコード可読性大幅向上
